+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Addded
+- SFTP support
+
+### Fixed
+- `.getInfos()` now correctly get the session
 
 ## [1.1.0] - 2017-04-26
 ### Added